refactor: split multimodule test files
